リファクタリング: PreviewModalコンポーネントを切り出し
